requirements.txt
annotated-types==0.7.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
fastai==2.7.18
fastcore==1.7.19
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
numpy<2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
opencv-python==4.10.0.84